requirements.txt
python-dotenv
requests
praw